+++ src/main/java/com/tbr/lettura/repository/ChallengeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tbr.lettura.model.Challenge;
 
-public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
     
 }
