+++ src/main/java/com/tbr/lettura/repository/UserChallengeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.tbr.lettura.model.UserChallenge;
 
 public interface UserChallengeRepository extends JpaRepository<UserChallenge, Long> {
-    List<UserChallenge> findByUserId(Long userId);
-    UserChallenge findByUserIdAndChallengeId(Long userId, Long challengeId);
+    List<UserChallenge> findByUserId(Integer userId);
+    UserChallenge findByUserIdAndChallengeId(Integer userId, Integer challengeId);
 }
